requirements.txt
# Self-Fortifying LLM Gateway Dependencies
# Python 3.10+ required

# Core ML/LLM libraries
dspy-ai>=2.2.0
mlflow>=2.11.0
openai>=1.12.0
anthropic>=0.18.0

# Visualization
matplotlib>=3.9.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Utilities
pyyaml>=6.0.1
rich>=13.7.0
click>=8.1.7

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# HTTP/API
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Development
python-dotenv>=1.0.0 